commit e5c421abb8bd517cb6964747b9ce23066e1a1cef
    9266 ixgbe: this statement may fall through
commit 8a7aa2a592747bd90d85645126d80688e024a0e2
    9673 move internet address manipulation functions to libc
commit 7e3488dc6cdcb0c04e1ce167a1a3bfef83b5f2e0
    9610 qede: error: left shift of negative value
commit d972451485b9ccf8c5c4951e2d24f63da9ac2cc8
    9609 qede: error: result of '983040 << 16' requires 37 bits to represent, but 'int' only has 32 bits
commit 3df3d01f9da8be5cdd1e81a5d050c24ccc55caa9
    9608 qede: error: 'free_buffer_count' may be used uninitialized in this function
commit 0d7de619fa7b28e48c8e80fa26d724400f4f47a4
    9607 qede: this statement may fall through
commit 77b62fe435d195b5a58f123e6c5ec9c9dffdc877
    9606 qede: error: this 'for' clause does not guard

	Conflicts:
	usr/src/man/man1m/route.1m
	usr/src/lib/libdladm/Makefile.com

+++ usr/src/uts/common/io/qede/579xx/drivers/ecore/ecore.h
@@ -111,7 +111,7 @@ enum ecore_nvm_cmd {
 
 #define SET_FIELD(value, name, flag)					\
 do {									\
-	(value) &= ~(name##_MASK << name##_SHIFT);			\
+	(value) &= ~((u64)name##_MASK << (u64)name##_SHIFT);		\
 	(value) |= ((((u64)flag) & (u64)name##_MASK) << (name##_SHIFT));\
 } while (0)
 
@@ -124,9 +124,10 @@ do {									\
 
 #define ECORE_MFW_SET_FIELD(name, field, value)				\
 do {									\
-	(name) &= ~((field ## _MASK) << (field ## _SHIFT));		\
+	(name) &= ~(((u64)field ## _MASK) << ((u64)field ## _SHIFT));	\
 	(name) |= (((value) << (field ## _SHIFT)) & (field ## _MASK));	\
 } while (0)
+
 
 static OSAL_INLINE u32 DB_ADDR(u32 cid, u32 DEMS)
 {
